Added contest providers
