+++ tvscheduler-react-app/tvscheduler-react-app/src/contexts/myShowsContext.js
@@ -2,7 +2,6 @@ import { createContext } from "react";
 
 const MyShowsContext = createContext({
   myShows: [],
-  addRemoveShow: () => {},
 });
 
 export default MyShowsContext;
